+++ docs/cookbooks/developers/content_lists.rst
@@ -1,0 +1,28 @@
+Content Lists
+=============
+
+Automatic Content Lists
+~~~~~~~~~~~~~~~~~~~~~~~
+
+Pinned articles
+---------------
+
+This feature gives you a way of exposing more important articles on top of the list.
+
+In automatic content lists you have the possibility to pin/unpin articles from the list. If you decide to pin one of the article from the list, it will always show up on top of the list, no matter how many new articles will be added to that list.
+
+You can pin as many articles as you want.
+
+Pinned article can be unpinned too. Once it's done, the unpinned article will be removed from the top of the list and will remain in the list on the corresponding position.
+
+By default, articles in Automatic Content List are ordered by "sticky" status (true or false)
+ and item creation date (desc).
+
+How articles are being added to Automatic Content Lists
+-------------------------------------------------------
+
+Once the new automatic content list is created and list's criteria are set for that list,
+articles which are published and match content list's criteria will be added automatically
+to Content List.
+
+If list's criteria are not defined, no articles will be added to the list.
