+++ src/SWP/Bundle/ContentListBundle/Resources/config/doctrine-orm/ContentList.orm.yml
