+++ src/SWP/Bundle/CoreBundle/DependencyInjection/SWPCoreExtension.php
@@ -14,13 +14,10 @@
 
 namespace SWP\Bundle\CoreBundle\DependencyInjection;
 
-use SWP\Bundle\CoreBundle\Model\ContentList;
-use SWP\Bundle\CoreBundle\Model\ContentListItem;
 use SWP\Bundle\StorageBundle\DependencyInjection\Extension\Extension;
 use SWP\Bundle\StorageBundle\Drivers;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Extension\PrependExtensionInterface;
 use Symfony\Component\DependencyInjection\Loader;
 
 /**
@@ -28,7 +25,7 @@ use Symfony\Component\DependencyInjection\Loader;
  *
  * To learn more see {@link http://symfony.com/doc/current/cookbook/bundles/extension.html}
  */
-class SWPCoreExtension extends Extension implements PrependExtensionInterface
+class SWPCoreExtension extends Extension
 {
     /**
      * {@inheritdoc}
@@ -48,33 +45,5 @@ class SWPCoreExtension extends Extension implements PrependExtensionInterface
         if ($config['device_listener']['enabled']) {
             $loader->load('device_listener.yml');
         }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function prepend(ContainerBuilder $container)
-    {
-        $config = $this->processConfiguration(new Configuration(), $container->getExtensionConfig($this->getAlias()));
-
-        if (!$container->hasExtension('swp_content_list')) {
-            return;
-        }
-
-        $container->prependExtensionConfig('swp_content_list', [
-            'persistence' => [
-                'orm' => [
-                    'enabled' => $config['persistence']['orm']['enabled'],
-                    'classes' => [
-                        'content_list' => [
-                            'model' => ContentList::class,
-                        ],
-                        'content_list_item' => [
-                            'model' => ContentListItem::class,
-                        ],
-                    ],
-                ],
-            ],
-        ]);
     }
 }
