+++ src/SWP/Bundle/CoreBundle/Resources/config/app/config.yml
@@ -98,6 +98,8 @@ swp_content_list:
             classes:
                 content_list:
                     model: SWP\Bundle\CoreBundle\Model\ContentList
+                content_list_item:
+                    model: SWP\Bundle\CoreBundle\Model\ContentListItem
                 list_content:
                     model: SWP\Bundle\CoreBundle\Model\Article
